+++ main.go
@@ -24,7 +24,7 @@ func main() {
 		BasePath:  "",
 		DstDirs:   dstDirs,
 		PartCount: partCount,
-		PartSize:  partSize,
+		PartSize:  int64(partSize),
 		ReDL:      map[FileState]bool{Completed: true, Resume: false, Broken: true},
 		UI:        ShowProgress,
 		IOMode: &IOMode{
